Add float support in compare_outputs, gdb_demo.sh and add gdb for pulp
This commmit automatically make a gdb_demo script based on the
compiled network.
This is necessary if the network outputs something different from
A regular int8 as output value
It also casts all read in values from a string to a float.
This is useful if the integer or float values need to be compared.

Furthermore this commit adds an option to use a locally installed
riscv gdb version, which can be used with remote debugging over
tcp:3333

+++ diana/byoc/compare_outputs.py
@@ -34,8 +34,8 @@ def get_gdb_output(gdb_log_path="debug/gdb.txt"):
             # makes a list of numbers in string format
             list_numbers = string.split(",")
             # convert strings to integers
-            values = [int(number) for number in list_numbers]
-        values_from_test = np.array(values, dtype="int8")
+            values = [float(number) for number in list_numbers]
+        values_from_test = np.array(values, dtype="float")
         # Values are returned from GDB in one big one-dimensional tensor
         # Reshaping here such that it matches the output
         return values_from_test
